Chore: 데이터베이스에서 사용하는 id 타입은 long으로 통일 (#28)

+++ src/main/java/com/hypeboy/hypeBoard/repository/CommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    Optional<Comment> findByPostId(int postId);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    Optional<Comment> findByPostId(Long postId);
 }
